requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai
pillow
pytesseract
python-dotenv
httpx
